+++ AddressBook/ERP_Core2/ScheduleEventsDomain/Repository/ScheduleEventRepository.cs
@@ -17,7 +17,7 @@ namespace MillenniumERP.ScheduleEventsDomain
         {
             _dbContext = (Entities)db;
         }
-        public async Task<IQueryable<ScheduleEvent>> GetScheduleEventsByEmployeeId(int employeeId)
+        public async Task<IQueryable<ScheduleEvent>> GetScheduleEventsByEmployeeId(long employeeId)
         {
             try
             {
